requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==4.2.20
exceptiongroup==1.2.2
feedparser==6.0.11
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==0.28.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
yt-dlp==2025.2.19
